feat: Modify Products component for consuming data
In this commit, the <Products /> component was modified such that it is
able to consume the data being provided by the ProductContext.Provider.
In doing this, the I imported the useContext hook and the ProductContext
within the <Products /> component. useContext is then passed the value
from the ProductContext i.e products and addItem respectively