Modify <ShoppingCart /> component consume Context data
In this commit, the <ShoppingCart /> component was modified such that it is
able to consume data from CartContext. In achieving this, useContext was
imported from the react library and CartContext was imported from CartContext.js

the data within CartContex is then retrieved and stored in cart variable by invoking
the useContext hook passing in CartContext.